Move DFM calc to link_frl_training
